requirements.txt
flask
requests
pymongo
python-dotenv
google-genai
duckduckgo-search
beautifulsoup4
playwright